\name{deepgmm-package}
\alias{deepgmm-package}
\alias{deepgmm}
\docType{package}
\title{
\packageTitle{deepgmm}
}
\description{
\packageDescription{deepgmm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{deepgmm}
\packageIndices{deepgmm}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{deepgmm}

Maintainer: \packageMaintainer{deepgmm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
