\name{deepgmm-package}
\alias{deepgmm-package}
\docType{package}
\title{
Deep Gaussian Mixture Models
\packageTitle{deepgmm}
}
\description{
The deep mixture model consists of a set of nested mixtures of
linear models, which globally provide a nonlinear model able
to describe the data in a very flexible way.
In order to avoid overparameterized solutions, dimension reduction
by factor models can be applied at each layer of the architecture,
thus resulting in deep mixtures of factor analyzers.
\packageDescription{deepgmm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{deepgmm}
\packageIndices{deepgmm}
}
\author{
\packageAuthor{deepgmm}

Maintainer: \packageMaintainer{deepgmm}
}
\references{
Viroli, C. and McLachlan, G.J. (2018).
Deep Gaussian mixture models. Statistics and Computing.
(Advance Access, published 1 December, 2017). To appear.
Preprint arXiv:1711.06929.
}
\keyword{package}
\keyword{clustering}
\keyword{multivariate}
\keyword{model}
% \seealso{
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }
% \examples{
% }
