\name{deepgmm-package}
\alias{deepgmm-package}
\docType{package}
\title{
\packageTitle{deepgmm}
}
\description{
\packageDescription{deepgmm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{deepgmm}
\packageIndices{deepgmm}
}
\author{
\packageAuthor{deepgmm}

Maintainer: \packageMaintainer{deepgmm}
}
\references{
Viroli, C. and McLachlan, G.J. (2018).
Deep Gaussian mixture models. Statistics and Computing.
(Advance Access, published 1 December, 2017). To appear.
Preprint arXiv:1711.06929.
}
\keyword{package}
\keyword{clustering}
\keyword{multivariate}
\keyword{model}
% \seealso{
% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }
% \examples{
% }
